Merging changes from AWS experiments

+++ tda/precision.py
@@ -1,3 +1,3 @@
 import torch
 
-default_tensor_type = torch.FloatTensor
+default_tensor_type = torch.DoubleTensor
